feat(client): Implement E2E reporting features and improve code organization
