+++ client/src/sections/e2e-page/components/TestResultsPerApp/constants.ts
@@ -1,0 +1,2 @@
+export const PAGE_SIZE = 10;
+
