+++ client/src/sections/e2e-page/index.ts
@@ -1,0 +1,3 @@
+import E2EPageContainer from '@/sections/e2e-page/E2EPageContainer.tsx';
+
+export default E2EPageContainer;
